+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added new post sharing options - contribution from @micahmo
 - Added Read All in inbox. - contribution from @ggriffo
 - Added support for lemmy 0.19.x authentication - contribution from @micahmo
+- Added support for accessibility profiles in settings - contribution from @micahmo
 
 ### Changed
 - Collapsed comments are easier to expand - contribution from @micahmo